+++ docs/examples/src/amazon_dynamodb_sessions_1.py
@@ -18,7 +18,7 @@ else:
 
 structure = Agent(
     conversation_memory=ConversationMemory(
-        driver=AmazonDynamoDbConversationMemoryDriver(
+        conversation_memory_driver=AmazonDynamoDbConversationMemoryDriver(
             session=boto3.Session(
                 aws_access_key_id=os.environ["AWS_ACCESS_KEY_ID"],
                 aws_secret_access_key=os.environ["AWS_SECRET_ACCESS_KEY"],