+++ griptape/configs/drivers/drivers_config.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING
 
 from attrs import define
 
@@ -13,6 +13,7 @@ from griptape.drivers import (
     DummyPromptDriver,
     DummyTextToSpeechDriver,
     DummyVectorStoreDriver,
+    LocalConversationMemoryDriver,
 )
 from griptape.utils.decorators import lazy_property
 
@@ -52,8 +53,8 @@ class DriversConfig(BaseDriversConfig):
         return DummyVectorStoreDriver(embedding_driver=self.embedding_driver)
 
     @lazy_property()
-    def conversation_memory_driver(self) -> Optional[BaseConversationMemoryDriver]:
-        return None
+    def conversation_memory_driver(self) -> BaseConversationMemoryDriver:
+        return LocalConversationMemoryDriver()
 
     @lazy_property()
     def text_to_speech_driver(self) -> BaseTextToSpeechDriver: